+++ integrations/formspree/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @gitbook/integration-formspree
 
+## 0.2.11
+
+### Patch Changes
+
+- 04be551: Update types in formspree integration to render email field correctly
+
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [5b37e83]
+    - @gitbook/runtime@0.21.0
+
 ## 0.2.9
 
 ### Patch Changes