+++ integrations/mermaid/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gitbook/integration-mermaid
 
+## 0.4.1
+
+### Patch Changes
+
+- 5cdad57: Fix mermaid diagram not rendering in Safari
+
+## 0.4.0
+
+### Minor Changes
+
+- bddebd8: Support ZenUML diagrams in Mermaid
+
 ## 0.3.4
 
 ### Patch Changes