+++ integrations/plausible/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @gitbook/integration-plausible
 
+## 0.8.0
+
+### Minor Changes
+
+- 772a6b3: Add cookies to plausible
+
+## 0.7.0
+
+### Minor Changes
+
+- 125bcb2: Move check for cookies into the tracking function
+
+## 0.6.0
+
+### Minor Changes
+
+- aac3f54: Add check to see if cookies should be disabled or not
+
+### Patch Changes
+
+- Updated dependencies [b72b44b]
+    - @gitbook/api@0.131.0
+
 ## 0.5.2
 
 ### Patch Changes