+++ integrations/reo/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @gitbook/integration-reo.dev
 
+## 0.5.0
+
+### Minor Changes
+
+- 125bcb2: Add cookie banner to tracking integrations
+
+## 0.4.0
+
+### Minor Changes
+
+- aac3f54: Add check to see if cookies should be disabled or not
+
+### Patch Changes
+
+- Updated dependencies [b72b44b]
+    - @gitbook/api@0.131.0
+
 ## 0.3.2
 
 ### Patch Changes