+++ integrations/slack/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @gitbook/integration-slack
 
+## 2.4.0
+
+### Minor Changes
+
+- 650e909: Update slack cache header
+
+## 2.3.0
+
+### Minor Changes
+
+- f531be5: Don't cache slack channels request
+
+## 2.2.0
+
+### Minor Changes
+
+- 1d4e22b: Figma and slack exposable error
+- 1d4e22b: Throw exposable error when slack is not configured
+
+## 2.1.0
+
+### Minor Changes
+
+- 8876593: Try to fix missing channels in Slack's configuration
+
 ## 2.0.4
 
 ### Patch Changes